refactor: features/friends 배럴 패턴 적용 및 import문 수정
