+++ src/features/friends/ui/index.ts
@@ -1,5 +1,6 @@
 export { default as AddFriendBottomSheet } from './add-friend-bottom-sheet'
 export { default as AddFriendButton } from './add-friend-button'
+export { default as DeleteFriendModal } from './delete-friend-modal'
 export { default as FriendItem } from './friend-item'
 export { default as FriendList } from './friend-list'
 export { default as FriendRequestButton } from './friend-request-button'
@@ -7,3 +8,5 @@ export { default as FriendRequestItem } from './friend-request-item'
 export { default as FriendRequestListBottomSheet } from './friend-request-list-bottom-sheet'
 export { default as FriendRequestListButton } from './friend-request-list-button'
 export { default as UserList } from './user-list'
+export { default as SearchTypeDropdown } from './search-type-dropdown'
+export { default as UserListSkeleton } from './user-list-skeleton'
